Update and add update-description files
1. Create UDFs for upcoming 4.x releases

The release process doc instructs us to do so for the next major version
(probably 4.5 as per the proposed release schedule) and its release candidate,
which was missed in the previous commit.

On top of that, while I'm at it, I'm also creating UDFs that advertise no
upgrade path at all for 4.3, 4.4, 4.6, 4.7, 4.8, and 4.9, in order to decrease
the impact in case the next RMs misunderstand this part of our doc. We already
have such UDFs for 4.1 and 4.2 so we're good there.

2. Update UDFs for all Tails 3.x released in the last year
   → manual upgrade to 4.0

The release process doc instructs us to do so for recent 3.x releases,
which was missed in the previous commit.

On top of that, it's useless for users to spend time automatically
upgrading an older 3.x to a more recent 3.x: they'll need to do a manual
upgrade to 4.0 anyway. So I'm also letting them know they should
directly upgrade to 4.0 manually.