Install virtualbox 6.0.12-dfsg-1 from our custom APT repository (refs: #17124, #17161).
sid now has 6.0.14-dfsg-1, and virtualbox-guest-x11 6.0.14-dfsg-1 depends on
a newer libc6 than the one we have in Buster. So for now, let's stick
to an older version, which seems preferable than fully removing this package.

A cleaner solution would be to maintain a custom backport of src:virtualbox.
Last time anonym and I discussed how much resources we wanted to spend into
VirtualBox support, we decided to do it on a best effort basis, i.e. to install
it from sid when it's possible, and to drop it when it's not possible anymore.
I'm already doing a bit more than that here, so maintaining a custom backport
seems unreasonably expensive to me.

We shall reconsider once we add Secure Boot support (the dkms module won't
be loadable anymore) or when VirtualBox has a severe security issue.