Test suite: don't have multiple instances of tcpdump writing to the same file, resulting in an unparsable network capture (refs: #17102)
The "Found something that is not an ethernet packet" test failure happens pretty
often in the "The MAC address is not leaked when booting Tails" scenario.
I suspect that's because during this scenario, we run two instances of Sniffer,
that have the same name, and therefore write to the same pcap file. They are
both started by the "I capture all network traffic" step: one in the Background,
the other one in this scenario itself. I would not be surprised if this
generated invalid pcap files, which is exactly the problem we see here.

Let's avoid this by dropping the Background, at the cost of duplicating
a few lines of Gherkin.

The other solutions I can think of don't look great:

 - We could move this scenario (that has the unique property here that we don't
   want to restore a snapshot: we want a complete boot) to a dedicated Feature.
   This makes no sense to me: this scenario really fits well into
   mac_spoofing.feature.

 - We could have "I capture all network traffic" deduplicate stuff behind the
   curtain before starting a new capture, i.e. detect whether another Sniffer
   instance with the same name is already running, and when that's the case, run
   @sniffer.stop and @sniffer.clear. I'm worried that the resulting code will be
   too complex and magical to be maintainable.