KeePassXC: don't list a non-existing file in the list of recent databases (refs: #17212)
With KeePassX, this trick made it so the the first time a user used KeePassX and
created their new database, the GUI proposed them, as the default destination,
a path under /home/amnesia/Persistent/ so the new database would persist.
But that does not work with KeePassXC, which our documentation already takes
into account. So we don't get any benefits anymore out of these
3 configuration lines.

OTOH, as reported by huertanix on #17212, this configuration hurts UX.