Let live-boot expose its /live/overlay as /lib/live/mount/overlay (refs: #15146)
/live/overlay (in the context of the initramfs) is the tmpfs
where the read-write branch of our union rootfs lives.

With aufs, this call to umount failed, and then live-boot would run:

   mount -o move /live/overlay /root/lib/live/mount/overlay

As a result, this tmpfs mount was visible outside of the initramfs,
and our initramfs-pre-shutdown-hook could unmount it on shutdown,
which ensured the data stored in there was cleaned from memory.

But with overlayfs, for some reason this call to umount succeeds, even though the
overlayfs upper layer (/live/overlay/rw) is stored in this filesystem, which
shows that this tmpfs is still mounted. As a result, this tmpfs is not
visible anymore, and cannot be unmounted on shutdown, so the data stored
in there remains in memory, available to cold-boot attackers.

Let's not unmount this tmpfs and go back to the same behavior we had
with aufs.

This will probably require bringing back some AppArmor-related automated
tests, that were removed on the #8415 branch precisely because live-boot
did not expose the overlay branch:

  12404eb883d8b68ab07242734d9a14a3d07f91ba
  c6541323cb70f7c2d1af77bb407bef0d3d3e554a
  a822c25f13c9673ffb39fb623e72c4d2894b112e
  4ee7e8d041a29d78e15a1bbca4b5dba1c1e09296