Report that overlayfs is not affected (Closes: #15689)
I'll also report on the ticket, but I want this comment to contain
"aufs" so it pops up when we grep for aufs in search of kludges
to clean up as part of #8415.

I'm not deleting this script yet because I've not thought about
what else we should do at the same time.