Endless upgrades: specify the IUK format version 2 (refs: #6876)
In our quest for (almost) endless incremental upgrades we'll end up
with larger IUKs than before. It turns out Tails Upgrader's memory
requirement of 2x the size of the IUK becomes a problem on systems
with 2 GB of memory, so we want to decrease the factor to 1x.

We achieve this by making the IUK mountable, so we can freely copy
contents from there onto the system partition without the need to
extract it *and* bundled archives, as was needed with the previous
format.

Based on commit 10950ddfe3611a0a8d89fd63fb68bbc9c63ee3a2
by anonym <anonym@riseup.net>.