Endless upgrades: switch to URL schema and UDF format version 2 (refs: #6876)
While drafting the deployment scenarios of #15281 for the next few release,
I realized that keeping the same URL schema and UDF format, while changing the
meaning of the version number in it, was too complicated to reason about, and
simply did not work in some cases. For example, when we release 4.3, we would
need UDFs under /upgrade/v1/Tails/4.0, that would be be used understood:

 - by the old Upgrader, as "that's what I want if I'm running 4.0"
   → these should point to IUKs in format v1 that upgrade to the latest
   release we provide IUKs v1 for (4.2)

 - by the new Upgrader, as "that's what I want if I was initially installed as
   4.0 and later upgraded to 4.2"
   → these should point to IUKs in format v2 that upgrade to 4.3

So let's avoid using the same URL schema with a component that can mean two very
different things depending on which Upgrader is asking; instead, let's give
ourselves the means to control independently what UDFs we feed versions of Tails
that have the old Upgrader and versions that have the new one.