Use a custom background image for GRUB (refs: #6560)
The syslinux boot menu is displayed in the center of the screen and in
the background image we use for syslinux, we put the Tails logo on the
left, so that it's displayed left of the boot menu.

The GRUB boot menu uses a bigger part of the screen. When using the same
background image as for syslinux, the logo looks out of place.

This adds a custom background image for GRUB, in which the logo is in
the center.