Thunderbird: disable automatic fetch of email at login and in the background (refs: #17222)
This replicates what Torbirdy did. It's not clear whether the benefits
of this behavior outweigh the problematic UX impact of not being notified
of new email automatically. We started discussing this on #17222 but
we have not reached a clear conclusion/consensus yet, so for now,
let's stick to the same behavior as previous versions of Tails,
and come back to this discussion later.