KeePassXC: Open ~/Persistent/keepassx.kdbx by default again (refs: #17212)
In commit 921022be214048a3caf8eabc1b643bae5387c2b0, we really only
wanted to remove the LastDatabases line, because that's the only option
that affects the list of last opened databases displayed to the user.

The LastOpenedDatabases option affects which databases KeePassXC tries
to open automatically.

The LastDir is used as the default directory in some file dialgos, for
example when choosing a keyfile.