Fix database renaming prompt not working with KeePassXC (refs: #17212)
KeePassXC creates "*.old.kdbx" backup files for each database file it
saves, and our code to show a dialog asking to rename the database only
opened the dialog if exactly one "*.kdbx" file is in the Persistent dir.

This commit ignores the "*.old.kdbx" files when counting the kdbx files
in the Persistent dir.