Explicit objective
The points in the "primary goals" are super vague on purpose. They come
from the aggregated goals that came out of the session at the 2017
summit and should stay like this. Search for "User goals for using
Tails" on /contribute/personas.

The fact that Riou could go to jail for their activities was already
implicit (cf. "Dissident bloggers and protesters have been jailed" and
"Achieve social change without being put under the spotlight") but it
was important to some of us to make it more explicit, according to some
feedback received at the summit 2018.