Test suite: close communication channels with child processes before killing them (refs: #14948)
The "close" method (https://ruby-doc.org/core-2.2.0/IO.html#method-i-close)
"Closes ios and flushes any pending writes to the operating system". This could
increase the chances that the child process we kill manages to exit (instead of
being left as a zombie), by closing any communication channels we may have left
open with it. But it's more likely that we get this expected benefit if we run
.close before we kill that child process, to increase the success chances of
whatever signal handler that child process has setup.