Upgrader: ensure debugging info lands in the Journal before we refer to it
While working on the test suite, it took me a while to figure out why the exit
code, stdout, and stderr of a failed command was not in the Journal. It turns
out that it's because our test suite does not properly close the zenity dialog
that we display before printing the more detailed debugging info to STDERR and
thus to the Journal, which makes debugging via the test suite impractical.

A user could be affected in a similar way: before this commit, when we told them
"For debugging information, execute the following command: sudo
tails-debugging-info", the info we were referring was actually not in the
Journal yet.