Test suite: avoid robustness step introducing fragility (refs: #15286)
I've seen cases when the ESC key stroke in "all notifications have disappeared"
closes the very dialog that we'll be looking for in the next step
("I am proposed to install an incremental upgrade to…"). I suspect
this never happened before because the Upgrader took longer to start,
so optimizing it introduced this race condition.

In these scenarios, I doubt that closing notifications is important: the
aforementioned zenity dialog is always displayed in the center of the screen and
there's little change a notification hides it.