IUK creation: only pass --xattrs to rsync when using overlayfs
On our Stretch isobuilders, passing --xattrs to rsync when generating an
aufs-based IUK makes it fail: setting xattrs on a file in an aufs filesystem
raises an "operation not supported" error.

We only need to preserve xattrs when building overlayfs-based IUKs,
because aufs will lose them at mount time anyway.