Jenkins: move design doc out of the blueprint
A blueprint is a great place to draft stuff while we develop it, but once it's
deployed in production, a non-blueprint, more stable place, is more suitable.
For example, it better reflects the fact this doc is about what we're actually
running, rather than about ideas we've had in the past and might have
implemented or not.