import-translations: work around the lack of usable branches in Tor's translation.git (refs: #17279)
For all our Transifex resources but tails-misc, Tor's translation.git currently
has no usable branch:

 - The _completed branches that we pull from only have languages which are fully
   translated and reviewed, i.e. very few translations.

 - The non-_completed branches have have non-reviewed translations, which we
   don't want to ship anymore.

The real fix will be https://redmine.tails.boum.org/code/issues/16774#note-25.
We've been waiting for it for 3 months now, and it's not sustainable
that the RM has to manually workaround the currently problematic situation
by hand during every release process.

So in the meantime, keep shipping the PO files we've already imported
when no suitable up-to-date one is available.