Release process: remove obsolete note (refs: #15287)
This problem has not happened since a few years, mostly because I'm now aware of
it and thus careful. Moreover, we now have a build_IUKs Jenkins job, that takes
a TAILS_GIT_COMMIT parameter, and can thus be used to test new IUK generation
code before it's merged and affects a RM, on real-world scenarios. I'm using
this regularly when working on the IUK codebase. As it happens, this job runs on
Jenkins' isobuilders, that currently run Stretch.

So all things considered, I'm now confident that this note does more
harm (scaring the RM) than good (providing useful and actionable information).