Vagrant build box: disable mitigation features for CPU vulnerabilities (refs: #17386)
Given the kind of things we do in our Vagrant build box, it seems very unlikely
that vulnerabilities such as Spectre and Meltdown can be exploited in there.
Let's reclaim some of the performance cost of the corresponding
mitigation features.

As of Linux 4.9.189, mitigations=off is equivalent to:

      nopti nospectre_v1 nospectre_v2 spectre_v2_user=off
      spec_store_bypass_disable=off l1tf=off mds=off

As of Linux 5.4.5, these extra options are added to the above list:

      kpti=0 nobp=0 ssbd=force-off tsx_async_abort=off kvm.nx_huge_pages=off