Build system: fully provision the Vagrant box every time it starts, and partially re-provision it for every build
Commit 5b2a3af4af533af8244f7c9a8e6cf028c18b7f0b said it would "provision a new
Vagrant VM for every ISO build" (except in some cases), but I'm not sure
I understand: we did exactly this before and that commit removed the very line
where we did run_vagrant('provision').

I understand the rest of that commit was supposed to ensure we provision the VM
"for every ISO build", but I've seen many weird failure modes in the last couple
years, on my system and coworkers', such as builds failing because build-tails
was not installed (which is supposed to have been done at provisioning time),
but Vagrant skipped the provisioning step because it was under the wrong
impression that the VM had been provisioned already¹. I suspect our mechanism to
make Vagrant forget that it did already provision a given box is not
fully reliable.

So, I'm bringing back an explicit provisioning step, this time in the vm:up
task. This should help guarantee that the provisioning was done before we start
the :build task.

Additionally, in order to mitigate the resulting performance regression for
subsequent builds when using the "rescue" or "keeprunning" build option, let's
skip unneeded provisioning steps when building for the second time (or more)
since the Vagrant box started: while it makes sense to run some of the
provisioning steps for every build, some other steps are not necessary for the
second build and the following ones.

[1] According to https://www.vagrantup.com/docs/provisioning/, provisioning is
    meant to be done when creating a VM for the first time, not every time
    Vagrant starts a VM. It looks like we're piggy-backing on a concept that
    means something else, so I would not be surprised if Vagrant did not do what
    we need by default.