IUK generation: don't make all files in the SquashFS diff owned by root (refs: #17422)
This fixes a regression introduced while implementing the IUK v2
format (iuk.git:b578872ba2167da7dbc106a9d20c88e523ab73b3).

We need to pass -all-root for the 1st-level SquashFS (the IUK itself)
but we need to *not* pass it for the SquashFS diff that's included
in the IUK, otherwise we're going to break systems when upgrading them.

Thanks kibi for independently doing the same root cause analysis,
and for testing that dropping -all-root indeed fixes the problem!