Upgrader: don't assume that all offered upgrades are relevant (refs: #17425)
Before this commit, the Upgrader blindly assumed that _any_ upgrade path that's
proposed for the initially installed version actually contains an upgrade to
a version that's strictly newer than the currently running one.

This assumption used to hold in the v1 scheme, but in the v2 scheme:

 - When running version == initially installed version, i.e. when one runs
   a Tails that was never upgraded, this assumption is true.

 - Else, this assumption is wrong.

As a consequence, any Tails >= 4.2 that hasn't this fix, and that ever got an
incremental upgrade applied, will keep proposing an upgrade even if it's
up-to-date already.

To fix this, let's filter the proposed upgrade paths to ignore those that would
"upgrade" to a lower-or-equal version than the currently running one.