website-cache: honor default values for command-line options
This reverts the "Use argparse's "required" parameter instead of checking if the
argument is None" part of 2f5785fde147dcc5eadba93f9f64f52ab4c638d8, which broke
the build when "cachewebsite" is passed:

   export WEBSITE_DEST_DIR=.
   WEBSITE_DEST_DIR=.
   export WEBSITE_CACHE_BASEDIR
   website-cache gc
   usage: website-cache [-h] --dest-dir DEST_DIR --cache-base-dir CACHE_BASE_DIR
                        [--debug]
                        {gc,get,put,key,forget} ...
   website-cache: error: the following arguments are required: --dest-dir, --cache-base-dir

For some reason, even if a default value is provided, an argument that has
required=True is required no matter what. The argparse documentation does not
specify what happens when default= and required= are both specified.