Test suite: load system load when using xdotool.
On my system I have seen a crazy loss of keystrokes: the step 'I start
Tails Installer' wants to type this into a terminal:

  /usr/bin/tails-installer --verbose > /tmp/tails-installer.log 2>&1

but I have seen it become just:

  /usr/bintaller.log 2>&1

so it seems a whole segment like

  /tails-installer --verbose > /tmp/tails-ins

was lost.

So let's see (temporarily!) log the load average before calling
xdotool to see if we can find a pattern. Perhaps we'll need to make
the system "settle" (e.g. wait until the load decreases) before
calling xdotool (or after doing heavy IO like copying images or saving
snapshots). I'm not sure load average is the best metric (I fear the
sliding windows of 1m/5m/15m are too long) so let's also look at
`vmstat` (although it's not ideal either).

Note also that this logging might affect the problem -- it might be
enough for the system to "settle" in some cases.