Test suite: make @screen.find() actually @screen.wait() for 2 seconds.
I suspect that Sikuli did this, and if so our test suite naturally
depends on it in many places. Alternatively, Sikuli is just much
slower than our current approach. The problem is made clear by this
example:

  When /^I create my XMPP account$/ do
    [...]
    @screen.click("PidginAddAccountProtocolLabel.png")
    @screen.click("PidginAddAccountProtocolXMPP.png")

The first click() opens a combobox and the second click() targets one
of its entries. I've seen the second click() fail because it was too
fast looking for PidginAddAccountProtocolXMPP.png -- the protocol
combobox had not even opened when the screenshot (which it will match
against) was taken!