Test suite: fix "Greeter disappears after restoring snapshot" workaround.
First of all, this problem does not occur on Jenkins (because it runs
Stretch and I run Sid/Buster?), and there I've seen the workaround
cause failures: the window is moved, but before the window is moved we
find the original position of the "Start Tails" button and the click
occurs after the window is moved, thus resulting in failure.

So let's try to fix this by adding --sync to the `xdotool windowmove`
call so it will block until the window has actually been
moved. Furthermore, let's skip the workaround entirely if it isn't
necessary (i.e. the Greeter is already visisble).