Test suite: handle non-English keyboards.
At the moment we only have a single test using a different locale than
English. In it we use German, and it so happens that the only keys we
have pressed so far are the same on both German and US English
keyboards. But this is not generally true (e.g. 'ä') so let's bail
when an unsupported key is requested.

Also, I had to switch @language from a class variable to a global one;
inside the Screen class we need to access @language, but it's not in
scope.