Test suite: Dogtail-ify pin entry handling.
Our OpenCV-based image matching is so fast that it sometimes causes
issues when we call maybe_deal_with_pinentry() twice in a
row (e.g. for symmetric encryption): when clicking OK in the first pin
entry we sometime manages to take the screenshot before the first pin
entry prompt has disappeared, so we start doing the things intended
for the second pin entry before it even has appeared, leading to
failure.