Test suite: fix issue where key modifiers affect subsequent key presses.
We have seen the test suite write "P?+(FJZ" when it was supposed to
type "P/+9fjz", just like if the Shift key got stuck at the "P" with
an US keyboard layout.

We try to fix this in two ways: First, by setting a holdtime > 0 when
telling libvirt to press keys. According to my tests, this alone might
fix it (even with delay = 0).

However, there's another issue: the libvirt's send_key() is
asynchronous, so it won't block until it has emitted the key, which
will get in the way when we write tests. For example:

    step 'I run "touch /tmp/test" in GNOME Terminal'
    assert($vm.file_exist?('/tmp/test')

can fail since the step invocation might return before it has typed
the command and pressed Return, so we'll check for the file before it
has been created. So, secondly, we enforce a big enough delay between
key presses to do this, *and* we move this delay from type() to
press() so we cannot escape this delay.