Requirements--
After reading quite a bit about recruitment strategies vs. diversity,
it seems to me that "you are good with" is too strong here, and may
discourage some people from even applying, in a way that negatively
affects the diversity demographics of applicants.

So let's move this to the role description, which in passing mostly resolves one
of the main concerns I had with the role description, i.e. it being both too
vague and relying a lot on a too detailed linked page.