Reimbursements: don't suggest that we'll need *all* receipts (refs: #16504)
The reviewer (hefee) pointed out that the way this was phrased previously
was stricter than our actual rules, which is confusing.

Also, it could lead people to not include in the sponsorship request some
expenses because they would be mistakenly lead to believe that they would not be
able to get them reimbursed.