Ensure we stop shipping RTL8822BE/RTL8822CE firmware once it's in Debian (refs: #17323)
Once those files are in Debian and they get updated there, if we don't notice
this warning, we'll be shipping outdated firmware, which is not ideal.

I'm concerned that this warning will be hard to spot in the middle of
a 15k-lines build log, that has 30+ other warnings which are all
false positives.

So let's abort the build whenever we detect this situation.