Release process: Clarify where IUKs come from.
I seem to remember anonym wondering where IUKs should be, and/or why
they would be there. Since I double-checked at the time, I'm not
wondering on my own, but I suppose it doesn't hurt being explicit about
the origin of the IUKs, and the need for waiting until the sync is over
(even if the non-reproducible IUKs case is supposed to be unlikely).