Add heart to Donate in top bar
As a way of making it a bit more eye-catching.

The heart in the "Free Software" image was well interpreted during the
usability tests:

P9:  Likes the heart better, made with love. Doesn't like always relating to
     money.

P10: Likes better with the heart than without because it carries more message
     of allowing people to escape the commercialization of data.

P12: Likes better the heart because it's not only about the price but also care
     and safety