Remove too dangerous option
How the most popular manufacturers instruct to upgrade the BIOS:

- https://support.lenovo.com/mx/en/solutions/ht500008
- https://support.hp.com/us-en/document/c00007682
- https://www.dell.com/support/article/es-mx/sln284433/what-is-bios-and-how-to-update-the-bios-on-your-dell-system?lang=en

Upgrading the BIOS can lead to a badly broken system if done erroneously
or if the procedure is interrupted.

I prefer not giving this option and loosing a few users than encouraging
people who might not actually need it to break their computer so badly.