Release process: catch up with master for real.
If switching to the master branch is needed, along with fetching, maybe
one should be merging back origin/master into master at the same time.
The following submodule update suggests this was the intent, so add the
“git merge” call. Also pass “--init” to “git submodule update” for good
measure.

Finally, be consistent and link all commands together with “&&”, instead
of keeping two separate command sequences in a single code block.