Release process: document post-release steps for RCs.
This is an initial attempt, which will tried for 4.5~rc1. I'm slightly
unsure if did the right thing with the two "thaw" instructions.

I opted to "fork" the steps for final releases, instead of adapting
them, to make things simpler. I'm sure we could make the same steps
work for both types of releases, but there would be a lot of "If you
just released a final release..." and so on, while an exhausted RM
likely would prefer more straightforward instructions.

Will-fix: #17376