Test suite: reminder to drop obsolete Fixnum (Will-fix: #17552).
In Ruby 2.4 the Fixnum (and Bignum) class was obsoleted, with Integer
as its replacement. In fact, in that version Fixnum == Integer,
explaining why this code works with >= 2.4 at all.

In Stretch we are stuck on Ruby 2.3.3, which has Integer but
Integer != Fixnum, so we're stuck with having to deal with the Fixnum
case until we drop support for Stretch.