tails-unblock-network: skip most graphics-related devices when triggering udev
On #17199 (and possibly #17228) I've seen breakage that seems to be caused by
udevadm trigger: the Greeter starts just fine, but on login, systemd-logind is
confused about the state of FDs (for /dev/dri/card{0,1}) it passes to X.Org:
when amnesia's X.Org starts, it gets passed paused FDs which makes it abort.
I suspect that's because at least one of /dev/dri/card{0,1} seem to briefly
disappear and come back again when we run udevadm trigger.

So let's try to avoid replaying kernel events for graphics-related devices.