Switch Japanese input method from Anthy to Mozc  (refs: #16719)
A Japanese speaker (passis12345678) tells us that:

"I don't think any Japanese users would agree that Anthy is better than Mozc.
The accuracy of Mozy's Kanji conversion is clearly superior to that of Anthy.
I’ve never seen or heard of anyone who likes Anthy.
Any Japanese user would be happy to replace Anthy with Mozc."

They also note that ibus-anthy has one advantage over ibus-mozc: it defaults to
Japanese input, while ibus-mozc defaults to alphabetic and the user has to
change the setting to Japanese every time they login. Despite that drawback,
passis12345678 thinks that switching to Mozc is worth it anyway: it's
a one-per-Tails-session annoyance, as opposed to a permanent annoyance while
typing in Japanese.

This would be fixed by fcitx-mozc (which defaults to Japanese), but switching to
fcitx is another, much bigger matter: IBus seems to be much better integrated
into GNOME than fcitx. Let's stick to IBus for now and try this minimal change.