GitLab: start listing repositories that should be migrated
This list was primarily built while working on
https://salsa.debian.org/tails-team/gitlab-migration/issues/30,
which was a good opportunity to identify repositories
that we link to.

If we don't migrate them all, we'll need to special-case the non-migrated ones
in tails.git as they're linked to via the tails_gitweb* ikiwiki shortcuts.
Doing that work would likely take more time than migrating the repos,
so let's do it.